// src/components/marketplaces/shopee/hooks/types.ts
export type Row = {
  [k: string]: any;
  ID: number;
  Loja: string;
  "ID Tray"?: string;
  "ID Var"?: string;
  Marca?: string;
  Referência?: string;
  Desconto?: number;
  Embalagem?: number;
  Frete?: number;
  Comissão?: number;
  Imposto?: number;
  Marketing?: number;
  "Margem de Lucro"?: number;
  Custo?: number;
  "Preço de Venda"?: number;
};